Layered Project
This still has some errors. Some variables does not exist in the current context
