+++ PostCreationDraft/Program.cs
@@ -1,343 +1,264 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.IO;
+using RULES;
+using MODELS;
+using System.Security.Principal;
 
-class Account
+
+namespace UI
 {
-	public string Username { get; set; }
-	public string Password { get; set; }
-	public string Name { get; set; }
-	public string StudentNumber { get; set; }
+    public class Program
+    {
+        static void Main(string[] args)
+        {
+            bool exit = false;
 
-	public Account(string username, string password, string name, string studentNumber)
-	{
-		Username = username;
-		Password = password;
-		Name = name;
-		StudentNumber = studentNumber;
-	}
-}
+            while (!exit)
+            {
+                if (loggedInUser == null)
+                {
+                    ShowLoginMenu();
+                    string option = Console.ReadLine();
+                    Console.WriteLine();
 
-class Post
-{
-	public int PostNumber { get; set; }
-	public string Content { get; set; }
-	public DateTime DateTimePosted { get; set; }
+                    switch (option)
+                    {
+                        case "1":
+                            Login();
+                            break;
+                        case "2":
+                            CreateAccount();
+                            break;
+                        case "3":
+                            exit = true;
+                            break;
+                        default:
+                            Console.WriteLine("Invalid option. Please try again.");
+                            break;
+                    }
+                }
+                else
+                {
+                    ShowMenu();
 
-	public Post(int postNumber, string content)
-	{
-		PostNumber = postNumber;
-		Content = content;
-		DateTimePosted = DateTime.Now;
-	}
-}
+                    string option = Console.ReadLine();
+                    Console.WriteLine();
 
-class Program
-{
-	static List<Account> accounts = new List<Account>();
-	static List<Post> posts = new List<Post>();
-	static string loggedInUser = null;
+                    switch (option)
+                    {
+                        case "1":
+                            CreatePost();
+                            break;
+                        case "2":
+                            EditPost();
+                            break;
+                        case "3":
+                            DeletePost();
+                            break;
+                        case "4":
+                            ListPosts();
+                            break;
+                        case "5":
+                            ShowAccountDetails();
+                            break;
+                        case "6":
+                            Logout();
+                            break;
+                        case "7":
+                            exit = true;
+                            break;
+                        default:
+                            Console.WriteLine("Invalid option. Please try again.");
+                            break;
+                    }
+                }
 
-	static void Main(string[] args)
-	{
+                Console.WriteLine();
+                Console.WriteLine("Press Enter to continue...");
+                Console.ReadLine();
+                Console.Clear();
+            }
 
-		bool exit = false;
+            void ShowLoginMenu()
+            {
+                Console.WriteLine("PUP HUB BULLETIN BOARD");
+                Console.WriteLine("1. Login");
+                Console.WriteLine("2. Create an account");
+                Console.WriteLine("3. Exit");
+                Console.Write("Select an option: ");
+            }
 
-		while (!exit)
-		{
-			if (loggedInUser == null)
-			{
-				ShowLoginMenu();
-				string option = Console.ReadLine();
-				Console.WriteLine();
+            void ShowMenu()
+            {
+                Console.WriteLine($"Welcome, {loggedInUser}!");
+                Console.WriteLine("1. Create a post");
+                Console.WriteLine("2. Edit a post");
+                Console.WriteLine("3. Delete a post");
+                Console.WriteLine("4. List all posts");
+                Console.WriteLine("5. Show account details");
+                Console.WriteLine("6. Logout");
+                Console.WriteLine("7. Exit");
+                Console.Write("Select an option: ");
+            }
 
-				switch (option)
-				{
-					case "1":
-						Login();
-						break;
-					case "2":
-						CreateAccount();
-						break;
-					case "3":
-						exit = true;
-						break;
-					default:
-						Console.WriteLine("Invalid option. Please try again.");
-						break;
-				}
-			}
-			else
-			{
-				ShowMenu();
+            void CreateAccount()
+            {
+                Console.WriteLine("Creating a new account");
 
-				string option = Console.ReadLine();
-				Console.WriteLine();
+                Console.Write("Enter a username: ");
+                string username = Console.ReadLine();
 
-				switch (option)
-				{
-					case "1":
-						CreatePost();
-						break;
-					case "2":
-						EditPost();
-						break;
-					case "3":
-						DeletePost();
-						break;
-					case "4":
-						ListPosts();
-						break;
-					case "5":
-						ShowAccountDetails();
-						break;
-					case "6":
-						Logout();
-						break;
-					case "7":
-						exit = true;
-						break;
-					default:
-						Console.WriteLine("Invalid option. Please try again.");
-						break;
-				}
-			}
+                if (IsUsernameTaken(username))
+                {
+                    Console.WriteLine("Username already taken. Please choose a different username.");
+                    return;
+                }
 
-			Console.WriteLine();
-			Console.WriteLine("Press Enter to continue...");
-			Console.ReadLine();
-			Console.Clear();
-		}
-	}
+                Console.Write("Enter a password: ");
+                string password = Console.ReadLine();
 
-	static void ShowLoginMenu()
-	{
-		Console.WriteLine("PUP HUB BULLETIN BOARD");
-		Console.WriteLine("1. Login");
-		Console.WriteLine("2. Create an account");
-		Console.WriteLine("3. Exit");
-		Console.Write("Select an option: ");
-	}
+                Console.Write("Enter your name(Last Name, First Name, MI): ");
+                string name = Console.ReadLine();
 
-	static void ShowMenu()
-	{
-		Console.WriteLine($"Welcome, {loggedInUser}!");
-		Console.WriteLine("1. Create a post");
-		Console.WriteLine("2. Edit a post");
-		Console.WriteLine("3. Delete a post");
-		Console.WriteLine("4. List all posts");
-		Console.WriteLine("5. Show account details");
-		Console.WriteLine("6. Logout");
-		Console.WriteLine("7. Exit");
-		Console.Write("Select an option: ");
-	}
+                Console.Write("Enter your student number: ");
+                string studentNumber = Console.ReadLine();
+                Account account = new Account(username, password, name, studentNumber);
+                accounts.Add(account);
 
-	static void CreateAccount()
-	{
-		Console.WriteLine("Creating a new account");
+                Console.WriteLine("Account created successfully!");
+                Console.WriteLine("Account Details:");
+                Console.WriteLine($"Username: {account.Username}");
+                Console.WriteLine($"Name: {account.Name}");
+                Console.WriteLine($"Student Number: {account.StudentNumber}");
+            }
 
-		Console.Write("Enter a username: ");
-		string username = Console.ReadLine();
+            void Login()
+            {
+                Console.Write("Enter your username: ");
+                string username = Console.ReadLine();
 
-		if (IsUsernameTaken(username))
-		{
-			Console.WriteLine("Username already taken. Please choose a different username.");
-			return;
-		}
+                Console.Write("Enter your password: ");
+                string password = Console.ReadLine();
 
-		Console.Write("Enter a password: ");
-		string password = Console.ReadLine();
+                Account account = GetAccount(username);
 
-		Console.Write("Enter your name(Last Name, First Name, MI): ");
-		string name = Console.ReadLine();
+                if (account != null && account.Password == password)
+                {
+                    loggedInUser = username;
+                    Console.WriteLine("Login successful!");
+                }
+                else
+                {
+                    Console.WriteLine("Invalid username or password.");
+                }
+            }
 
-		Console.Write("Enter your student number: ");
-		string studentNumber = Console.ReadLine();
+            void Logout()
+            {
+                loggedInUser = null;
+                Console.WriteLine("Logged out successfully!");
+            }
 
-		Account account = new Account(username, password, name, studentNumber);
-		accounts.Add(account);
+            void CreatePost()
+            {
+                Console.WriteLine("Creating a new post");
 
-		Console.WriteLine("Account created successfully!");
-		Console.WriteLine("Account Details:");
-		Console.WriteLine($"Username: {account.Username}");
-		Console.WriteLine($"Name: {account.Name}");
-		Console.WriteLine($"Student Number: {account.StudentNumber}");
-	}
+                Console.Write("Enter post content: ");
+                string content = Console.ReadLine();
 
-	static bool IsUsernameTaken(string username)
-	{
-		foreach (var account in accounts)
-		{
-			if (account.Username == username)
-			{
-				return true;
-			}
-		}
+                int postNumber = GetNextPostNumber();
 
-		return false;
-	}
+                Post post = new Post(postNumber, content);
+                posts.Add(post);
 
-	static void Login()
-	{
-		Console.Write("Enter your username: ");
-		string username = Console.ReadLine();
+                Console.WriteLine("Post created successfully!");
+            }
 
-		Console.Write("Enter your password: ");
-		string password = Console.ReadLine();
+            void EditPost()
+            {
+                Console.WriteLine("Editing a post");
 
-		Account account = GetAccount(username);
+                ListPosts();
 
-		if (account != null && account.Password == password)
-		{
-			loggedInUser = username;
-			Console.WriteLine("Login successful!");
-		}
-		else
-		{
-			Console.WriteLine("Invalid username or password.");
-		}
-	}
+                Console.Write("Enter the number of the post to edit: ");
+                int postNumber = int.Parse(Console.ReadLine());
 
-	static Account GetAccount(string username)
-	{
-		foreach (var account in accounts)
-		{
-			if (account.Username == username)
-			{
-				return account;
-			}
-		}
+                Post post = GetPostByNumber(postNumber);
 
-		return null;
-	}
+                if (post != null)
+                {
+                    Console.Write("Enter new post content: ");
+                    string newContent = Console.ReadLine();
 
-	static void Logout()
-	{
-		loggedInUser = null;
-		Console.WriteLine("Logged out successfully!");
-	}
+                    post.Content = newContent;
 
-	static void CreatePost()
-	{
-		Console.WriteLine("Creating a new post");
+                    Console.WriteLine("Post edited successfully!");
+                }
+                else
+                {
+                    Console.WriteLine("Post not found.");
+                }
+            }
 
-		Console.Write("Enter post content: ");
-		string content = Console.ReadLine();
+            void DeletePost()
+            {
+                Console.WriteLine("Deleting a post");
 
-		int postNumber = GetNextPostNumber();
+                ListPosts();
 
-		Post post = new Post(postNumber, content);
-		posts.Add(post);
+                Console.Write("Enter the number of the post to delete: ");
+                int postNumber = int.Parse(Console.ReadLine());
 
-		Console.WriteLine("Post created successfully!");
-	}
+                Post post = GetPostByNumber(postNumber);
 
-	static int GetNextPostNumber()
-	{
-		if (posts.Count > 0)
-		{
-			return posts[posts.Count - 1].PostNumber + 1;
-		}
+                if (post != null)
+                {
+                    posts.Remove(post);
 
-		return 1;
-	}
+                    Console.WriteLine("Post deleted successfully!");
+                }
+                else
+                {
+                    Console.WriteLine("Post not found.");
+                }
+            }
 
-	static void EditPost()
-	{
-		Console.WriteLine("Editing a post");
+            void ListPosts()
+            {
+                Console.WriteLine("List of posts:");
 
-		ListPosts();
+                if (posts.Count == 0)
+                {
+                    Console.WriteLine("No posts available.");
+                }
+                else
+                {
+                    foreach (var post in posts)
+                    {
+                        Console.WriteLine($"Post  {post.PostNumber}");
+                        Console.WriteLine($"Content: {post.Content}");
+                        Console.WriteLine($"Posted on: {post.DateTimePosted}");
+                        Console.WriteLine();
+                    }
+                }
+            }
 
-		Console.Write("Enter the number of the post to edit: ");
-		int postNumber = int.Parse(Console.ReadLine());
+            void ShowAccountDetails()
+            {
+                Account account = GetAccount(loggedInUser);
 
-		Post post = GetPostByNumber(postNumber);
+                if (account != null)
+                {
+                    Console.WriteLine("Account Details:");
+                    Console.WriteLine($"Username: {account.Username}");
+                    Console.WriteLine($"Name: {account.Name}");
+                    Console.WriteLine($"Student Number: {account.StudentNumber}");
+                }
+                else
+                {
+                    Console.WriteLine("Account not found.");
+                }
+            }
 
-		if (post != null)
-		{
-			Console.Write("Enter new post content: ");
-			string newContent = Console.ReadLine();
-
-			post.Content = newContent;
-
-			Console.WriteLine("Post edited successfully!");
-		}
-		else
-		{
-			Console.WriteLine("Post not found.");
-		}
-	}
-
-	static void DeletePost()
-	{
-		Console.WriteLine("Deleting a post");
-
-		ListPosts();
-
-		Console.Write("Enter the number of the post to delete: ");
-		int postNumber = int.Parse(Console.ReadLine());
-
-		Post post = GetPostByNumber(postNumber);
-
-		if (post != null)
-		{
-			posts.Remove(post);
-
-			Console.WriteLine("Post deleted successfully!");
-		}
-		else
-		{
-			Console.WriteLine("Post not found.");
-		}
-	}
-
-	static void ListPosts()
-	{
-		Console.WriteLine("List of posts:");
-
-		if (posts.Count == 0)
-		{
-			Console.WriteLine("No posts available.");
-		}
-		else
-		{
-			foreach (var post in posts)
-			{
-				Console.WriteLine($"Post  {post.PostNumber}");
-				Console.WriteLine($"Content: {post.Content}");
-				Console.WriteLine($"Posted on: {post.DateTimePosted}");
-				Console.WriteLine();
-			}
-		}
-	}
-
-	static Post GetPostByNumber(int postNumber)
-	{
-		foreach (var post in posts)
-		{
-			if (post.PostNumber == postNumber)
-			{
-				return post;
-			}
-		}
-
-		return null;
-	}
-
-	static void ShowAccountDetails()
-	{
-		Account account = GetAccount(loggedInUser);
-
-		if (account != null)
-		{
-			Console.WriteLine("Account Details:");
-			Console.WriteLine($"Username: {account.Username}");
-			Console.WriteLine($"Name: {account.Name}");
-			Console.WriteLine($"Student Number: {account.StudentNumber}");
-		}
-		else
-		{
-			Console.WriteLine("Account not found.");
-		}
-	}
+        }
+    }
 }
