+++ RULES/GetAccountRule.cs
@@ -1,0 +1,22 @@
+﻿using System;
+using DataLayer;
+using MODELS;
+
+namespace RULES
+{
+    public class GetAccountRule
+    {
+        static Account GetAccount(string username)
+        {
+            foreach (var account in accounts)
+            {
+                if (account.Username == username)
+                {
+                    return account;
+                }
+            }
+
+            return null;
+        }
+    }
+}
