+++ RULES/GetPostByNumberRule.cs
@@ -1,0 +1,22 @@
+﻿using System;
+using DataLayer;
+using MODELS;
+
+namespace RULES
+{
+    public class GetPostbyNumberRule
+    {
+        static Post GetPostByNumber(int postNumber)
+        {
+            foreach (var post in posts)
+            {
+                if (post.PostNumber == postNumber)
+                {
+                    return post;
+                }
+            }
+
+            return null;
+        }
+    }
+}
